Refactor connection from display for photo list components
